requirements.txt
pywin32
urllib3
iso8601
Flask_Cors
requests
pypresence
valclient
pystray
pyperclip
cursor
psutil
Flask
InquirerPy
Pillow
PyInstaller
websockets
pyyaml
pyshortcuts
winshell 
win10toast
black
isort
mypy
pylint
ruff
nest_asyncio
lcu-driver
win32gui
win32process